Attempt at getting the ball rolling on Billing/Plans

+++ pkg/model/payment.go
@@ -7,6 +7,6 @@ import (
 type Payment struct {
 	InvoiceDate time.Time `json:"invoicedate"`
 	PaymentDate time.Time `json:"paymentdate"`
-	Amount      float32   `json:"amount"`
+	Amount      float64   `json:"amount"`
 	Currency    string    `json:"currency"`
 }
